AnomolyDetectionServer/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
pycparser==2.21
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ripe.atlas.cousteau==1.4.2
ripe.atlas.sagan==1.3.1
six==1.16.0
socketIO-client==0.7.2
sqlparse==0.4.2
urllib3==1.26.7
websocket-client==0.59.0
